Moving exception into categorie
