+++ cogs/archiveuse/profile_manager.py
@@ -5,13 +5,13 @@ from libs.config import Config
 from libs.databases.user import User
 from libs.databases.wallpaper import Wallpaper
 from libs.databases.wallpapers import Wallpapers
-from libs.exception.color_not_correct_exception import ColorNotCorrectException
-from libs.exception.not_enougt_smartpoint_exception import NotEnougtsmartpointException
-from libs.exception.unable_to_download_wallpaper_exception import UnableToDownloadImageException
-from libs.exception.wallpaper_already_posseded_exception import WallpaperAlreadyPossededException
-from libs.exception.wallpaper_cannot_be_buyed_exception import WallpaperCannotBeBuyedException
-from libs.exception.wallpaper_not_exist_exception import WallpaperNotExistException
-from libs.exception.wallpaper_not_posseded_exception import WallpaperNotPossededException
+from libs.exception.color.color_not_correct_exception import ColorNotCorrectException
+from libs.exception.smartpoint.not_enougt_smartpoint_exception import NotEnougtsmartpointException
+from libs.exception.wallpaper.unable_to_download_wallpaper_exception import UnableToDownloadImageException
+from libs.exception.wallpaper.wallpaper_already_posseded_exception import WallpaperAlreadyPossededException
+from libs.exception.wallpaper.wallpaper_cannot_be_buyed_exception import WallpaperCannotBeBuyedException
+from libs.exception.wallpaper.wallpaper_not_exist_exception import WallpaperNotExistException
+from libs.exception.wallpaper.wallpaper_not_posseded_exception import WallpaperNotPossededException
 from libs.log import Log, LogType
 from libs.paginator import Paginator
 from libs.utils import Utils
