+++ libs/databases/user.py
@@ -4,12 +4,12 @@ from libs.databases.database_access_implement import DatabaseAccessImplement, Pr
 from libs.databases.sqlite.sqlite_access import SqliteAccess
 from libs.databases.wallpaper import Wallpaper
 from libs.databases.wallpapers import Wallpapers
-from libs.exception.color_not_correct_exception import ColorNotCorrectException
-from libs.exception.not_enougt_smartpoint_exception import NotEnougtsmartpointException
-from libs.exception.wallpaper_already_posseded_exception import WallpaperAlreadyPossededException
-from libs.exception.wallpaper_cannot_be_buyed_exception import WallpaperCannotBeBuyedException
-from libs.exception.wallpaper_not_exist_exception import WallpaperNotExistException
-from libs.exception.wallpaper_not_posseded_exception import WallpaperNotPossededException
+from libs.exception.color.color_not_correct_exception import ColorNotCorrectException
+from libs.exception.smartpoint.not_enougt_smartpoint_exception import NotEnougtsmartpointException
+from libs.exception.wallpaper.wallpaper_already_posseded_exception import WallpaperAlreadyPossededException
+from libs.exception.wallpaper.wallpaper_cannot_be_buyed_exception import WallpaperCannotBeBuyedException
+from libs.exception.wallpaper.wallpaper_not_exist_exception import WallpaperNotExistException
+from libs.exception.wallpaper.wallpaper_not_posseded_exception import WallpaperNotPossededException
 from libs.utils import Utils
 
 class User:
