+++ libs/databases/wallpaper.py
@@ -1,6 +1,6 @@
 from libs.databases.database_access_implement import DatabaseAccessImplement
 from libs.databases.sqlite.sqlite_access import SqliteAccess
-from libs.exception.wallpaper_not_exist_exception import WallpaperNotExistException
+from libs.exception.wallpaper.wallpaper_not_exist_exception import WallpaperNotExistException
 
 
 class Wallpaper:
