+++ libs/exception/badge/badge_not_exist_exception.py
