+++ libs/exception/smartpoint/not_enougt_smartpoint_exception.py
