+++ libs/exception/wallpaper/wallpaper_already_posseded_exception.py
