+++ tests/test_wallpapers.py
@@ -2,7 +2,7 @@ import unittest
 
 from libs.databases.wallpaper import Wallpaper
 from libs.databases.wallpapers import Wallpapers
-from libs.exception.wallpaper_already_exist_exception import WallpaperAlreadyExistException
+from libs.exception.wallpaper.wallpaper_already_exist_exception import WallpaperAlreadyExistException
 
 class TestWallpapers(unittest.TestCase):
     __wallpaper : Wallpapers
